requirements.txt
django
djangorestframework
pandas
geopy
psycopg2-binary